+++ service/src/main/resources/db/migration/V362__message_thread_sensitivity_flag.sql
@@ -1,0 +1,3 @@
+ALTER TABLE message_draft ADD COLUMN sensitive boolean default false not null;
+ALTER TABLE message_thread ADD COLUMN sensitive boolean default false not null;
+ALTER TABLE message_thread ALTER COLUMN sensitive DROP DEFAULT;
